+++ lib/oneview-sdk/resource.rb
@@ -29,13 +29,13 @@ module OneviewSDK
     # @param [Integer] api_ver The api version to use when interracting with this resource.
     #   Defaults to the client.api_version if it exists, or the OneviewSDK::Client::DEFAULT_API_VERSION.
     def initialize(client, params = {}, api_ver = nil)
-      fail InvalidClient, 'Must specify a valid client' unless client.is_a?(OneviewSDK::Client)
+      raise InvalidClient, 'Must specify a valid client' unless client.is_a?(OneviewSDK::Client)
       @client = client
       @logger = @client.logger
       @api_version = api_ver || @client.api_version
       if @api_version > @client.max_api_version
-        fail UnsupportedVersion,
-             "#{self.class.name} api_version '#{@api_version}' is greater than the client's max_api_version '#{@client.max_api_version}'"
+        raise UnsupportedVersion,
+              "#{self.class.name} api_version '#{@api_version}' is greater than the client's max_api_version '#{@client.max_api_version}'"
       end
       @data ||= {}
       set_all(params)
@@ -45,7 +45,7 @@ module OneviewSDK
     # @note Name or URI must be specified inside the resource
     # @return [Boolean] Whether or not retrieve was successful
     def retrieve!
-      fail IncompleteResource, 'Must set resource name or uri before trying to retrieve!' unless @data['name'] || @data['uri']
+      raise IncompleteResource, 'Must set resource name or uri before trying to retrieve!' unless @data['name'] || @data['uri']
       results = self.class.find_by(@client, name: @data['name']) if @data['name']
       results = self.class.find_by(@client, uri:  @data['uri'])  if @data['uri'] && (!results || results.empty?)
       return false unless results.size == 1
@@ -57,7 +57,7 @@ module OneviewSDK
     # @note name or uri must be specified inside resource
     # @return [Boolean] Whether or not resource exists
     def exists?
-      fail IncompleteResource, 'Must set resource name or uri before trying to retrieve!' unless @data['name'] || @data['uri']
+      raise IncompleteResource, 'Must set resource name or uri before trying to retrieve!' unless @data['name'] || @data['uri']
       return true if @data['name'] && self.class.find_by(@client, name: @data['name']).size == 1
       return true if @data['uri']  && self.class.find_by(@client, uri:  @data['uri']).size == 1
       false
@@ -206,7 +206,7 @@ module OneviewSDK
       when :yml, :yaml
         File.open(file_path, 'w') { |f| f.write(temp_data.to_yaml) }
       else
-        fail InvalidFormat, "Invalid format: #{format}"
+        raise InvalidFormat, "Invalid format: #{format}"
       end
       true
     end
@@ -296,19 +296,19 @@ module OneviewSDK
 
     # Fail unless @client is set for this resource.
     def ensure_client
-      fail IncompleteResource, 'Please set client attribute before interacting with this resource' unless @client
+      raise IncompleteResource, 'Please set client attribute before interacting with this resource' unless @client
       true
     end
 
     # Fail unless @data['uri'] is set for this resource.
     def ensure_uri
-      fail IncompleteResource, 'Please set uri attribute before interacting with this resource' unless @data['uri']
+      raise IncompleteResource, 'Please set uri attribute before interacting with this resource' unless @data['uri']
       true
     end
 
     # Fail for methods that are not available for one resource
     def unavailable_method
-      fail MethodUnavailable, "The method ##{caller[0][/`.*'/][1..-2]} is unavailable for this resource"
+      raise MethodUnavailable, "The method ##{caller[0][/`.*'/][1..-2]} is unavailable for this resource"
     end
 
     private
@@ -316,7 +316,7 @@ module OneviewSDK
     # Recursive helper method for like?
     # Allows comparison of nested hash structures
     def recursive_like?(other, data = @data)
-      fail "Can't compare with object type: #{other.class}! Must respond_to :each" unless other.respond_to?(:each)
+      raise "Can't compare with object type: #{other.class}! Must respond_to :each" unless other.respond_to?(:each)
       other.each do |key, val|
         return false unless data && data.respond_to?(:[])
         if val.is_a?(Hash)
@@ -332,22 +332,18 @@ module OneviewSDK
 
   # Get resource class that matches the type given
   # @param [String] type Name of the desired class type
+  # @param [Fixnum] api_ver API module version to fetch resource from
   # @return [Class] Resource class or nil if not found
-  def self.resource_named(type)
-    classes = {}
-    orig_classes = []
-    OneviewSDK.constants.each do |c|
-      klass = OneviewSDK.const_get(c)
-      next unless klass.is_a?(Class) && klass < OneviewSDK::Resource
-      name = klass.name.split('::').last
-      orig_classes.push(name)
-      classes[name.downcase.delete('_').delete('-')] = klass
-      classes["#{name.downcase.delete('_').delete('-')}s"] = klass
-    end
+  def self.resource_named(type, api_ver = @api_version)
+    raise "API version #{api_ver} is not supported!" unless SUPPORTED_API_VERSIONS.include?(api_ver)
     new_type = type.to_s.downcase.gsub(/[ -_]/, '')
-    return classes[new_type] if classes.keys.include?(new_type)
+    api_module = OneviewSDK.const_get("API#{api_ver}")
+    api_module.constants.each do |c|
+      klass = api_module.const_get(c)
+      next unless klass.is_a?(Class) && klass < OneviewSDK::Resource
+      name = klass.name.split('::').last.downcase.delete('_').delete('-')
+      return klass if new_type =~ /^#{name}[s]?$/
+    end
+    nil
   end
 end
-
-# Load all resources:
-Dir[File.dirname(__FILE__) + '/resource/*.rb'].each { |file| require file }